Factor out methods common to API examples
